Refactored test classes and updated client creation methods
Test classes are relocated and renamed to better match their use cases. ClientEntityTestObjectMother and ClientTestObjectMother's method parameters are also updated. Annotations for 'CreateClientRequest' and 'CreateClientResponse' in 'ClientRestAdapter' are added for the builder pattern. The 'addClient' method is renamed to 'createClient' for clarity.
