+++ client-back/src/main/java/com/gml/infrastructure/adapters/inbound/rest/response/CreateClientResponse.java
@@ -1,7 +1,10 @@
 package com.gml.infrastructure.adapters.inbound.rest.response;
 
+import lombok.Builder;
+
 import java.util.Date;
 
+@Builder
 public record CreateClientResponse (
         String sharedKey,
         String bussinessId,
