+++ client-back/src/test/java/com/gml/domain/implementation/testobjectmother/ClientEntityTestObjectMother.java
@@ -1,4 +1,4 @@
-package com.gml.persistence.testobjectmother;
+package com.gml.domain.implementation.testobjectmother;
 
 
 import com.gml.infrastructure.adapters.outbound.persistence.entity.ClientEntity;
@@ -6,10 +6,10 @@ import com.gml.infrastructure.adapters.outbound.persistence.entity.ClientEntity;
 public class ClientEntityTestObjectMother {
     public static ClientEntity aClientEntity() {
         return ClientEntity.builder()
-                .sharedKey("1L")
+                .sharedKey("gsoto")
                 .bussinessId("name")
                 .email("description@gmail.com")
-                .phone("453453454")
+                .phone("34567776")
                 .build();
     }
 }
